# backend/openapi.yml
openapi: 3.0.3
info:
  title: DocIntel API Documentation
  description: API specifications for the DocIntel document management system. Includes authentication, folder management, file uploading/downloading, and sharing permissions.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: /
    description: Current host server
security:
  - BearerAuth: []
paths:
...
  /api/documents/upload/initiate:
    post:
      summary: Initiate single-part or multipart document upload
      tags:
        - Documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadInitiateRequest'
      responses:
        '200':
          description: Upload instructions returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadInitiateResponse'
  /api/documents/{id}/complete:
    post:
      summary: Complete single-part or multipart document upload
      tags:
        - Documents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadCompleteRequest'
      responses:
        '200':
          description: Upload marked complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
  /api/documents/resolve-path:
    post:
      summary: Dynamic path creation and resolver (Finder utility)
      tags:
        - Folders
      parameters:
        - in: query
          name: relativePath
          required: true
          schema:
            type: string
          description: Path hierarchy like "SubfolderA/SubfolderB"
        - in: query
          name: parentFolderId
          required: false
          schema:
            type: string
            format: uuid
          description: Initial directory ID
      responses:
        '200':
          description: Path resolved and folders created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderResponse'
        '403':
          description: Insufficient permissions on the parent folder
...
  /api/documents/{id}/presigned-url:
    get:
      summary: Get an AWS S3 pre-signed download URL for a document
      tags:
        - Documents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Pre-signed URL generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
        '403':
          description: Insufficient permissions
  /api/documents/download/{id}:
    get:
      summary: Direct document download link (302 Redirect to pre-signed S3 URL)
      tags:
        - Documents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '302':
          description: Redirecting to S3 pre-signed download URL
  /api/documents/view/{id}:
    get:
      summary: Direct inline document preview link (302 Redirect to pre-signed S3 URL)
      tags:
        - Documents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '302':
          description: Redirecting to S3 pre-signed preview URL
  /api/documents/{id}:
    delete:
      summary: Delete a document
      tags:
        - Documents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document deleted successfully
        '403':
          description: Insufficient permissions
  /api/documents/folders/{id}:
    delete:
      summary: Delete a folder and recursively all its contents
      tags:
        - Folders
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Folder and all contents deleted successfully
        '403':
          description: Insufficient permissions
  /api/documents/folders/download-zip/{id}:
    get:
      summary: Download all folder contents zipped
      tags:
        - Folders
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: ZIP file binary stream containing the folder contents
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '403':
          description: Insufficient permissions
  /api/documents/search:
    get:
      summary: Paginated search and filtering of documents
      tags:
        - File Explorer
      parameters:
        - in: query
          name: search
          required: false
          schema:
            type: string
          description: Case-insensitive search keyword
        - in: query
          name: category
          required: false
          schema:
            type: string
          description: Document category enum string
        - in: query
          name: favorite
          required: false
          schema:
            type: boolean
          description: Filter for starred/favorite documents only
        - in: query
          name: tag
          required: false
          schema:
            type: string
          description: Filter by document tag
        - in: query
          name: page
          required: false
          schema:
            type: integer
            default: 0
          description: Page index (0-based)
        - in: query
          name: size
          required: false
          schema:
            type: integer
            default: 10
          description: Items per page
      responses:
        '200':
          description: Paginated documents list
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/DocumentResponse'
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
  /api/documents/{id}/favorite:
    put:
      summary: Toggle favorite/starred flag for a document
      tags:
        - Documents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Starred status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
  /api/documents/{id}/tags:
    put:
      summary: Update tags string on a document
      tags:
        - Documents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
              example: "work, invoice, important"
      responses:
        '200':
          description: Tags updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
  /api/documents/move/{id}:
    put:
      summary: Move a document to a new folder or the root directory
      tags:
        - Documents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the document to move
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveDocumentRequest'
      responses:
        '200':
          description: Document moved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '403':
          description: Insufficient permissions on document or destination folder
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MoveDocumentRequest:
      type: object
      properties:
        folderId:
          type: string
          format: uuid
          nullable: true
          description: ID of the folder, or null to move to root directory.
    FolderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        parentId:
          type: string
          format: uuid
          nullable: true
...
    UploadInitiateRequest:
      type: object
      required:
        - name
        - size
        - contentType
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
        folderId:
          type: string
          format: uuid
          nullable: true
        category:
          type: string
          nullable: true
    UploadInitiateResponse:
      type: object
      properties:
        documentId:
          type: string
          format: uuid
        uploadId:
          type: string
          nullable: true
        urls:
          type: array
          items:
            type: string
            format: uri
        chunkSize:
          type: integer
    UploadCompleteRequest:
      type: object
      properties:
        uploadId:
          type: string
          nullable: true
        parts:
          type: array
          items:
            $ref: '#/components/schemas/PartDTO'
    PartDTO:
      type: object
      properties:
        partNumber:
          type: integer
        eTag:
          type: string
          nullable: true
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
        status:
          type: string
        folderId:
          type: string
          format: uuid
          nullable: true
        category:
          type: string
          nullable: true
        favorite:
          type: boolean
        tags:
          type: string
...